Extract shared breakdown bar view
